// apps/mutils-web/src/api/model/collectionEntrySpherePerks.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Shiron.Mutils.API | v1
 * OpenAPI spec version: 1.0.0
 */

export interface CollectionEntrySpherePerks {
	/** @pattern ^-?(?:0|[1-9]\d*)$ */
	perk1: number | string;
	/** @pattern ^-?(?:0|[1-9]\d*)$ */
	perk2: number | string;
	/** @pattern ^-?(?:0|[1-9]\d*)$ */
	perk3: number | string;
	/** @pattern ^-?(?:0|[1-9]\d*)$ */
	perk4: number | string;
	/** @pattern ^-?(?:0|[1-9]\d*)$ */
	perk5: number | string;
	perk6: boolean;
	perk7: boolean;
	perk8: boolean;
	perk9: boolean;
	perk10: boolean;
}